Refactor utility functions to be tested into their own file
