+++ tools/mdbook_preprocessor/mdbook_preprocessor.js
@@ -1,6 +1,8 @@
 // @ts-check
 /// <reference types="node" />
 
+import { tokenize, extractCommand } from "./util.js";
+
 /**
  * @typedef {import("./types").Book} Book
  * @typedef {import("./types").BookItem} BookItem
@@ -217,71 +219,3 @@ function processCodeWrap(code, indent) {
       .join("")
   }</pre></code>`;
 }
-
-/**
- * This function tokenizes a line of text into strings that should be kept
- * together when wrapping text. For example, the text "[-p | --prototype]"
- * should be a single token.
- *
- * @example
- * tokenize("git town append [-p | --prototype] <branch-name>")
- * // => ["git", "town", "append", "[-p | --prototype]", "<branch-name>"]
- *
- * @param {string} line
- * @returns {string[]}
- */
-function tokenize(line) {
-  const GROUP_CHARS = ["()", "<>", "[]"];
-
-  const tokens = [];
-  let token = "";
-  let group = undefined;
-  for (const char of line) {
-    if (group) {
-      if (char === group[1]) {
-        group = undefined;
-      }
-      token += char;
-    } else {
-      const nextGroup = GROUP_CHARS.find(group => group[0] === char);
-      if (nextGroup) {
-        group = nextGroup;
-        token += char;
-      } else if (char === " ") {
-        tokens.push(token);
-        token = "";
-      } else {
-        token += char;
-      }
-    }
-  }
-  tokens.push(token);
-  return tokens;
-}
-
-/**
- * This function extracts the command and other tokens from a line of text.
- *
- * @example
- * extractCommand(["git", "town", "append", "[-p | --prototype]", "<branch-name>"])
- * // => { command: "git town append", otherTokens: ["[-p | --prototype]", "<branch-name>"] }
- *
- * @param {string[]} tokens
- * @returns {{ command: string, otherTokens: string[] }}
- */
-function extractCommand(tokens) {
-  const otherTokens = [...tokens];
-
-  const commandTokens = [];
-  while (otherTokens.length > 0 && otherTokens[0].match(/^[a-z]/i)) {
-    commandTokens.push(otherTokens[0]);
-    otherTokens.shift();
-  }
-
-  const command = commandTokens.join(" ");
-
-  return { command, otherTokens };
-}
-
-// Make TypeScript think this file is a module
-export {};
